+++ static/files/leaders.json
@@ -1,1 +1,1 @@
-[{"name": "DaBoss", "date": "2018-07-09 07:29:37.102172", "CPM": 23.61}, {"name": "DaBoss", "date": "2018-06-15 16:29:27.058039", "CPM": 22.77}, {"name": "Mr21", "date": "2018-08-01 07:31:59.691276", "CPM": 21.64}, {"name": "Mr19", "date": "2018-08-01 07:25:48.423658", "CPM": 19.32}, {"name": "James", "date": "2018-07-09 20:43:59.943085", "CPM": 16.51}, {"name": "James", "date": "2018-07-09 20:43:59.943085", "CPM": 16.51}, {"name": "James", "date": "2018-07-09 20:43:59.943085", "CPM": 16.51}, {"name": "James", "date": "2018-07-09 20:30:26.945827", "CPM": 15.83}, {"name": "Captain Chaos", "date": "2018-07-09 07:29:37.102172", "CPM": 14.53}, {"name": "DaBoss", "date": "2018-05-25 16:49:21.792532", "CPM": 13.6}]
+[{"name": "DaBoss", "date": "2018-07-09 07:29:37.102172", "CPM": 23.61}, {"name": "DaBoss", "date": "2018-06-15 16:29:27.058039", "CPM": 22.77}, {"name": "Mr21", "date": "2018-08-01 07:31:59.691276", "CPM": 21.64}, {"name": "Twenty", "date": "2018-08-17 16:30:35.094164", "CPM": 20.2}, {"name": "Jackeroo", "date": "2018-08-13 21:06:33.854976", "CPM": 19.35}, {"name": "Mr19", "date": "2018-08-01 07:25:48.423658", "CPM": 19.32}, {"name": "James", "date": "2018-07-09 20:43:59.943085", "CPM": 16.51}, {"name": "James", "date": "2018-07-09 20:43:59.943085", "CPM": 16.51}, {"name": "James", "date": "2018-07-09 20:43:59.943085", "CPM": 16.51}, {"name": "James", "date": "2018-07-09 20:30:26.945827", "CPM": 15.83}]
