+++ static/files/leaders20.json
@@ -1,1 +1,1 @@
-[{"name": "again", "date": "2018-07-27 16:32:37.728462", "CPM": 0.0}, {"name": "DBoss", "date": "2018-05-29 07:56:41.725360", "CPM": 2.24}, {"name": "DBoss", "date": "2018-05-29 17:30:46.178474", "CPM": 2.43}, {"name": "60%dude", "date": "2018-07-27 16:32:37.728462", "CPM": 10.63}, {"name": "OtherJack", "date": "2018-08-01 07:15:33.371007", "CPM": 21.25}]
+[{"name": "OtherJack", "date": "2018-08-01 07:15:33.371007", "CPM": 21.25}, {"name": "Nineteen", "date": "2018-08-17 16:30:35.094164", "CPM": 19.28}, {"name": "60%dude", "date": "2018-07-27 16:32:37.728462", "CPM": 10.63}, {"name": "DBoss", "date": "2018-05-29 17:30:46.178474", "CPM": 2.43}, {"name": "DBoss", "date": "2018-05-29 07:56:41.725360", "CPM": 2.24}, {"name": "again", "date": "2018-07-27 16:32:37.728462", "CPM": 0.0}]
