+++ templates/index.html
@@ -3,12 +3,21 @@
     <section class="jumbotron text-center">
         <div class="container">
           <h1 class="jumbotron-heading">Multiplication Ninja</h1>
-            <p class="lead text-muted">Use this game to improve your skills with times tables </br>The faster and more accurate you get the better.</br></br>
-            You can play this game as a set number of questions (20) </br>or play for 1 minute and see how many you can complete!</p>
+            <p class="lead text-muted">Use the games to improve your skills with times tables </br>The faster and more accurate you get the better.</br></br>
+            You can play the games as a set number of questions (20) </br>or play for 1 minute and see how many you can complete!</p>
           <p>
-            <a href="{{url_for('start_game',game_mode='multi_set')}}" class="btn btn-primary my-2">Start 20 question game</a>
-            <a href="{{url_for('start_game',game_mode='multi_time')}}" class="btn btn-primary my-2">Start 1 minute game</a>
+            <table class="table table-borderless">
+            <tbody><tr>
+                <td>Straight Multiplication</td>
+                <td><a href="{{url_for('start_game',game_mode='multi_set')}}" class="btn btn-primary my-2">Start 20 question game
+                    </a></td>
+                <td><a href="{{url_for('show_lb_direct',game_mode='multi_set')}}"><span class="fas fa-list fa-3x"></span> </a></td>
+                <td><a href="{{url_for('start_game',game_mode='multi_time')}}" class="btn btn-primary my-2">Start 1 minute game</a></td>
+                <td><a href="{{url_for('show_lb_direct',game_mode='multi_time')}}"><span class="fas fa-list fa-3x"></span> </a></td>
+            </tr>
 
+            </tbody>
+        </table>
           </p>
         </div>
     </section>
