--- templates/index_old.html
@@ -1,106 +1,0 @@
-<!doctype html>
-<html class="no-js" lang="">
-    <head>
-        <meta charset="utf-8">
-        <meta http-equiv="x-ua-compatible" content="ie=edge">
-        <!-- Bootstrap CSS -->
-        <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
-
-        <title>Multi Ninja</title>
-        <meta name="description" content="">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-
-        <link rel="manifest" href="site.webmanifest">
-        <link rel="apple-touch-icon" href="icon.png">
-        <!-- Place favicon.ico in the root directory -->
-
-        <link rel="stylesheet" href="/static/css/normalize.css">
-        <link rel="stylesheet" href="/static/css/main.css">
-        <link rel="stylesheet" href="/static/css/game.css" >
-    </head>
-    <body>
-        <!--[if lte IE 9]>
-            <p class="browserupgrade">You are using an <strong>outdated</strong> browser. Please <a href="https://browsehappy.com/">upgrade your browser</a> to improve your experience and security.</p>
-        <![endif]-->
-
-        <!-- Add your site or application content here -->
-
-        <header>
-              <div class="collapse bg-dark" id="navbarHeader">
-                <div class="container">
-                  <div class="row">
-                    <div class="col-sm-8 col-md-7 py-4">
-                      <h4 class="text-white">About</h4>
-                      <p class="text-muted">Improve your skills at multiplication. Become the Master</p>
-                    </div>
-                    <div class="col-sm-4 offset-md-1 py-4">
-                      <h4 class="text-white">Contact</h4>
-                      <ul class="list-unstyled">
-                        <li><a href="#" class="text-white">Follow on Twitter</a></li>
-                        <li><a href="#" class="text-white">Like on Facebook</a></li>
-                        <li><a href="#" class="text-white">Email me</a></li>
-                      </ul>
-                    </div>
-                  </div>
-                </div>
-              </div>
-              <div class="navbar navbar-dark bg-dark box-shadow">
-                <div class="container d-flex justify-content-between">
-                  <a href="#" class="navbar-brand d-flex align-items-center">
-                    <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="mr-2"><path d="M23 19a2 2 0 0 1-2 2H3a2 2 0 0 1-2-2V8a2 2 0 0 1 2-2h4l2-3h6l2 3h4a2 2 0 0 1 2 2z"></path><circle cx="12" cy="13" r="4"></circle></svg>
-                    <strong>Multiplication Ninja</strong>
-                  </a>
-                  <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarHeader" aria-controls="navbarHeader" aria-expanded="false" aria-label="Toggle navigation">
-                    <span class="navbar-toggler-icon"></span>
-                  </button>
-                </div>
-              </div>
-        </header>
-
-        <main>
-            <section class="jumbotron text-center">
-                    <div class="container">
-                      <h1 class="jumbotron-heading">Multiplication Ninja</h1>
-                        <p class="lead text-muted">Use this game to improve your skills with times tables </br>The faster and more accurate you get the better.</br></br>
-                        You can play this game as a set number of questions (20) </br>or play for 1 minute and see how many you can complete!</p>
-                      <p>
-                        <a href="{{url_for('start_game',game_mode='Set Q')}}" class="btn btn-primary my-2">Start 20 question game</a>
-                        <a href="{{url_for('start_game',game_mode='Set time')}}" class="btn btn-primary my-2">Start 1 minute game</a>
-
-                      </p>
-                    </div>
-            </section>
-
-
-        </main>
-
-        <footer class="text-muted">
-              <div class="container">
-                <p class="float-right">
-                  <a href="#">Back to top</a>
-                </p>
-                  <p>Multiplication Ninja is &copy; A# Systems & Consulting</p>
-                  <p><a href = "http://www.asharpsystems.com"> Visit our page</a></p>
-              </div>
-        </footer>
-
-
-         <!-- jQuery first, then Popper.js, then Bootstrap JS -->
-        <script src="/static/js/modernizr-3.5.0.min.js"></script>
-        <script src="https://code.jquery.com/jquery-3.2.1.min.js" integrity="sha256-hwg4gsxgFZhOsEEamdOYGBf13FyQuiTwlAQgxVSNgt4=" crossorigin="anonymous"></script>
-        <script>window.jQuery || document.write('<script src="static/js/jquery-3.2.1.min.js"><\/script>')</script>
-        <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js" integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q" crossorigin="anonymous"></script>
-        <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js" integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl" crossorigin="anonymous"></script>
-
-        <script src="/static/js/plugins.js"></script>
-        <script src="/static/js/main.js"></script>
-        <script src="/static/js/game_1.js"></script>
-        <!-- Google Analytics: change UA-XXXXX-Y to be your site's ID.
-        <script>
-            window.ga=function(){ga.q.push(arguments)};ga.q=[];ga.l=+new Date;
-            ga('create','UA-XXXXX-Y','auto');ga('send','pageview')
-        </script> -->
-        <!--<script src="https://www.google-analytics.com/analytics.js" async defer></script>-->
-
-    </body>
-</html>
